BE/requirements.txt
# FastAPI Quiz Assistant Server Dependencies
# Core requirements for the current implementation

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP and CORS
httpx==0.25.2

# Data validation and parsing
pydantic==2.5.0

# Environment variables
python-dotenv==1.0.0

# OpenAI API client
openai==1.3.7

# Async support
asyncio-throttle==1.0.2

# Future enhancements (commented for now, uncomment when needed)
# Authentication & Security
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-multipart==0.0.6

# Database support (PostgreSQL)
# asyncpg==0.29.0
# sqlalchemy==2.0.23
# alembic==1.12.1

# Additional utilities
# redis==5.0.1
# celery==5.3.4

# Development and testing
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0
